# web-bff/src/swaggerOutput.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0.0",
    "title": "Manage users",
    "description": "FreshLink manage users service"
  },
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": ""
    }
  ],
  "paths": {
...
    "/orders/": {
      "get": {
        "tags": [
          "Orders"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "No token": {
                    "value": {
                      "message": "No token provided"
                    }
                  },
                  "JWT payload": {
                    "value": {
                      "message": "Cannot parse JWT payload"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Bad token": {
                    "value": {
                      "message": "Invalid or expired token"
                    }
                  },
                  "Role": {
                    "value": {
                      "message": "Your role does not have access"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully sent request for further processing"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "No token": {
                    "value": {
                      "message": "No token provided"
                    }
                  },
                  "JWT payload": {
                    "value": {
                      "message": "Cannot parse JWT payload"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Bad token": {
                    "value": {
                      "message": "Invalid or expired token"
                    }
                  },
                  "Role": {
                    "value": {
                      "message": "Your role does not have access"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "_id": "67f9899c5be73585d33d44ae",
                "restaurant_id": "restaurant12345",
                "farmer_id": "farmer456",
                "products": [
                  {
                    "product_id": "apple00123",
                    "quantity": 105.5
                  },
                  {
                    "product_id": "banana002",
                    "quantity": 7.2
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/orders/allInfo": {
      "get": {
        "tags": [
          "Orders"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "No token": {
                    "value": {
                      "message": "No token provided"
                    }
                  },
                  "JWT payload": {
                    "value": {
                      "message": "Cannot parse JWT payload"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Bad token": {
                    "value": {
                      "message": "Invalid or expired token"
                    }
                  },
                  "Role": {
                    "value": {
                      "message": "Your role does not have access"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Order ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "No token": {
                    "value": {
                      "message": "No token provided"
                    }
                  },
                  "JWT payload": {
                    "value": {
                      "message": "Cannot parse JWT payload"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Bad token": {
                    "value": {
                      "message": "Invalid or expired token"
                    }
                  },
                  "Role": {
                    "value": {
                      "message": "Your role does not have access"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Orders"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Order ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully sent request for further processing"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "No token": {
                    "value": {
                      "message": "No token provided"
                    }
                  },
                  "JWT payload": {
                    "value": {
                      "message": "Cannot parse JWT payload"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Bad token": {
                    "value": {
                      "message": "Invalid or expired token"
                    }
                  },
                  "Role": {
                    "value": {
                      "message": "Your role does not have access"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "_id": "67f9899c5be73585d33d44ae",
                "restaurant_id": "restaurant12345",
                "farmer_id": "farmer456",
                "products": [
                  {
                    "product_id": "apple00123",
                    "quantity": 105.5
                  },
                  {
                    "product_id": "banana002",
                    "quantity": 7.2
                  }
                ]
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Order ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully sent request for further processing"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "No token": {
                    "value": {
                      "message": "No token provided"
                    }
                  },
                  "JWT payload": {
                    "value": {
                      "message": "Cannot parse JWT payload"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Bad token": {
                    "value": {
                      "message": "Invalid or expired token"
                    }
                  },
                  "Role": {
                    "value": {
                      "message": "Your role does not have access"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/farmers/": {
      "get": {
        "tags": [
          "Farmers"
        ],
        "summary": "Get all farmers",
        "description": "Returns a list of all farmers",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with a list of farmers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/farmer"
                  }
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/arrayOfFarmers"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "No token": {
                    "value": {
                      "message": "No token provided"
                    }
                  },
                  "JWT payload": {
                    "value": {
                      "message": "Cannot parse JWT payload"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Bad token": {
                    "value": {
                      "message": "Invalid or expired token"
                    }
                  },
                  "Role": {
                    "value": {
                      "message": "Your role does not have access"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/error500"
                  }
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Farmers"
        ],
        "summary": "Register a new farmer",
        "description": "Adds a new farmer",
        "responses": {
          "200": {
            "description": "Farmer created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/farmer"
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/farmer"
                  }
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Already exists": {
                    "value": {
                      "message": "Email already exists"
                    }
                  }
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Missing required properties": {
                    "value": {
                      "message": "All required properties must be defined"
                    }
                  }
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/error500"
                  }
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/hiddenSchemas/postFarmer"
              },
              "examples": {
                "example1": {
                  "$ref": "#/components/examples/postFarmer"
                }
              }
            }
          }
        }
      }
    },
...
    "/restaurants/": {
      "get": {
        "tags": [
          "Restaurants"
        ],
        "summary": "Get all restaurants",
        "description": "Returns a list of all restaurants",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with a list of restaurants",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/restaurant"
                  }
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/arrayOfRestaurants"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "No token": {
                    "value": {
                      "message": "No token provided"
                    }
                  },
                  "JWT payload": {
                    "value": {
                      "message": "Cannot parse JWT payload"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Bad token": {
                    "value": {
                      "message": "Invalid or expired token"
                    }
                  },
                  "Role": {
                    "value": {
                      "message": "Your role does not have access"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/error500"
                  }
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Restaurants"
        ],
        "summary": "Register a new restaurant",
        "description": "Adds a new restaurant",
        "responses": {
          "200": {
            "description": "Restaurant created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/restaurant"
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/restaurant"
                  }
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Already exists": {
                    "value": {
                      "message": "Email already exists"
                    }
                  }
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Missing required properties": {
                    "value": {
                      "message": "All required properties must be defined"
                    }
                  }
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/error500"
                  }
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/hiddenSchemas/postRestaurant"
              },
              "examples": {
                "example1": {
                  "$ref": "#/components/examples/postRestaurant"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Get the token from login endpoints"
      }
    },
    "schemas": {
...
      "farmer": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "677164acd5343f0dac83716a"
                  }
                }
              },
              "email": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "john.doe@example.com"
                  }
                }
              },
              "farmName": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "Doe's Farm"
                  }
                }
              },
              "description": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "A nice farm"
                  }
                }
              },
              "name": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "John"
                  }
                }
              },
              "surname": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "Doe"
                  }
                }
              },
              "location": {
                "$ref": "#/components/schemas/location"
              }
            }
          }
        }
      },
...
      "error": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "message": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "Example error message"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "examples": {
...
      "arrayOfFarmers": {
        "summary": "A sample array of farmer objects",
        "value": [
          {
            "_id": "677164acd5343f0dac83716a",
            "email": "john.doe@example.com",
            "farmName": "Doe's Farm",
            "description": "A nice farm.",
            "name": "John",
            "surname": "Doe",
            "location": {
              "place": "3124 MontvaleDr, Springfield",
              "postCode": "62704 IL",
              "country": "USA"
            }
          }
        ]
      },
...
      "error500": {
        "summary": "Server error",
        "value": {
          "message": "Internal Server Error"
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}